+++ src/axom/mint/tests/mint_mesh_particle_mesh.cpp
@@ -11,7 +11,7 @@
 
 // Sidre includes
 #ifdef AXOM_MINT_USE_SIDRE
-  #include "axom/sidre/core/sidre.hpp"
+  #include "axom/sidre.hpp"
 namespace sidre = axom::sidre;
 #endif