+++ src/axom/spin/tests/spin_implicit_grid.cpp
@@ -459,10 +459,14 @@ TYPED_TEST(ImplicitGridExecTest, get_candidates_pt_vectorized)
   using BBox = typename TestFixture::BBox;
   using GridT = typename TestFixture::GridT;
   using SpacePt = typename TestFixture::SpacePt;
+  using ExecSpace = typename TestFixture::ExecSpace;
 
   SLIC_INFO("Test ImplicitGrid getCandidatesAsArray() with "
-            << axom::execution_space<typename TestFixture::ExecSpace>::name()
+            << axom::execution_space<ExecSpace>::name()
             << " execution space for points in " << DIM << "D");
+
+  int kernelAllocID = axom::execution_space<ExecSpace>::allocatorID();
+  int hostAllocID = axom::execution_space<axom::SEQ_EXEC>::allocatorID();
 
   // Note: A 10 x 10 x 10 implicit grid in the unit cube.
   //       Grid cells have a spacing of .1 along each dimension
@@ -494,9 +498,29 @@ TYPED_TEST(ImplicitGridExecTest, get_candidates_pt_vectorized)
     };
 
     axom::Array<int> count, offset, candidates;
+    {
+      // Copy query points to the device
+      axom::ArrayView<const SpacePt> queryPtsHost(queryPts, 9);
+      axom::Array<SpacePt> queryPtsDevice(queryPtsHost, kernelAllocID);
 
-    // Run query against implicit grid
-    grid.getCandidatesAsArray(9, queryPts, offset, count, candidates);
+      axom::Array<int> countDevice(9, 9, kernelAllocID);
+      axom::Array<int> offsetDevice(9, 9, kernelAllocID);
+      axom::Array<int> candidatesDevice;
+
+      // Run query against implicit grid
+      grid.getCandidatesAsArray(9,
+                                queryPtsDevice.data(),
+                                offsetDevice,
+                                countDevice,
+                                candidatesDevice);
+
+      // Copy results back to the host
+      count = axom::Array<int>(countDevice, hostAllocID);
+      offset = axom::Array<int>(offsetDevice, hostAllocID);
+      candidates = axom::Array<int>(candidatesDevice, hostAllocID);
+    }
+
+    // Copy results back to the host
 
     // Test some points that are expected to match
     for(int i = 0; i < 5; ++i)
@@ -528,9 +552,26 @@ TYPED_TEST(ImplicitGridExecTest, get_candidates_pt_vectorized)
                           SpacePt {.85, .85, .85}};
 
     axom::Array<int> count, offset, candidates;
+    {
+      // Copy query points to the device
+      axom::ArrayView<const SpacePt> queryPtsHost(queryPts, 3);
+      axom::Array<SpacePt> queryPtsDevice(queryPtsHost, kernelAllocID);
 
-    // Run query against implicit grid
-    grid.getCandidatesAsArray(3, queryPts, offset, count, candidates);
+      axom::Array<int> countDevice(3, 3, kernelAllocID);
+      axom::Array<int> offsetDevice(3, 3, kernelAllocID);
+      axom::Array<int> candidatesDevice;
+      // Run query against implicit grid
+      grid.getCandidatesAsArray(3,
+                                queryPtsDevice.data(),
+                                offsetDevice,
+                                countDevice,
+                                candidatesDevice);
+
+      // Copy results back to the host
+      count = axom::Array<int>(countDevice, hostAllocID);
+      offset = axom::Array<int>(offsetDevice, hostAllocID);
+      candidates = axom::Array<int>(candidatesDevice, hostAllocID);
+    }
     std::unordered_set<int> expected[] = {{2}, {3}, {2, 3}};
 
     for(int i = 0; i < 3; i++)
@@ -729,10 +770,14 @@ TYPED_TEST(ImplicitGridExecTest, get_candidates_box_vectorized)
   using BBox = typename TestFixture::BBox;
   using GridT = typename TestFixture::GridT;
   using SpacePt = typename TestFixture::SpacePt;
+  using ExecSpace = typename TestFixture::ExecSpace;
 
   SLIC_INFO("Test ImplicitGrid getCandidatesAsArray() with "
-            << axom::execution_space<typename TestFixture::ExecSpace>::name()
+            << axom::execution_space<ExecSpace>::name()
             << " execution space for boxes in " << DIM << "D");
+
+  int kernelAllocID = axom::execution_space<ExecSpace>::allocatorID();
+  int hostAllocID = axom::execution_space<axom::SEQ_EXEC>::allocatorID();
 
   // Note: A 10 x 10 x 10 implicit grid in the unit cube.
   //       Grid cells have a spacing of .1 along each dimension
@@ -825,9 +870,27 @@ TYPED_TEST(ImplicitGridExecTest, get_candidates_box_vectorized)
                                    DIM * 10};
 
   axom::Array<int> offset, count, candidates;
+  {
+    // Copy query points to the device
+    axom::ArrayView<const BBox> queryBoxesHost(queryBoxes, N_QUERIES);
+    axom::Array<BBox> queryBoxesDevice(queryBoxesHost, kernelAllocID);
 
-  // Run query against implicit grid
-  grid.getCandidatesAsArray(N_QUERIES, queryBoxes, offset, count, candidates);
+    axom::Array<int> countDevice(N_QUERIES, N_QUERIES, kernelAllocID);
+    axom::Array<int> offsetDevice(N_QUERIES, N_QUERIES, kernelAllocID);
+    axom::Array<int> candidatesDevice;
+
+    // Run query against implicit grid
+    grid.getCandidatesAsArray(N_QUERIES,
+                              queryBoxesDevice.data(),
+                              offsetDevice,
+                              countDevice,
+                              candidatesDevice);
+
+    // Copy results back to the host
+    count = axom::Array<int>(countDevice, hostAllocID);
+    offset = axom::Array<int>(offsetDevice, hostAllocID);
+    candidates = axom::Array<int>(candidatesDevice, hostAllocID);
+  }
 
   for(int i = 0; i < N_QUERIES; i++)
   {